pyproject.toml
[project]
name = "multiple-auth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.8"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "pyotp>=2.9.0",
    "python-dotenv>=1.1.0",
    "qrcode[pil]>=8.2",
]
